rtp: discard packets when unable to match with any media
When packets come with an unknown SSRC, it's not possible to know where
to send them. The RtpBin will have audio and/or video input sink pads,
but we need to be able to know which one is the correct one.

If previous identification methods have failed at this point, we still
don't know where to send the buffers. But the src pad cannot be left
unlinked.

Previously, this would cause a GStreamer stream error "reason: not
linked". Packets would stop flowing from the upstream source (the
NiceSrc pad that receives packets from the network). Now, at least that
error is avoided, and the whole flow is not interrupted.